+++ dcmdata/libsrc/dcostrmz.cc
@@ -21,10 +21,10 @@
  *
  *  Purpose: zlib compression filter for output streams
  *
- *  Last Update:      $Author: meichel $
- *  Update Date:      $Date: 2002-08-29 15:57:49 $
+ *  Last Update:      $Author: wilkens $
+ *  Update Date:      $Date: 2002-12-20 14:55:34 $
  *  Source File:      $Source: /export/gitmirror/dcmtk-git/../dcmtk-cvs/dcmtk/dcmdata/libsrc/dcostrmz.cc,v $
- *  CVS/RCS Revision: $Revision: 1.2 $
+ *  CVS/RCS Revision: $Revision: 1.3 $
  *  Status:           $State: Exp $
  *
  *  CVS/RCS Log at end of file
@@ -177,7 +177,7 @@ Uint32 DcmZLibOutputFilter::fillInputBuffer(const void *buf, Uint32 buflen)
       result = DCMZLIBOUTPUTFILTER_BUFSIZE - (inputBufStart_ + inputBufCount_);
       if (result > buflen) result = buflen;
 
-      memcpy(inputBuf_ + inputBufStart_ + inputBufCount_, data, result);
+      memcpy(inputBuf_ + inputBufStart_ + inputBufCount_, data, (size_t)result);
       inputBufCount_ += result;
       data += result;
       buflen -= result;
@@ -190,7 +190,7 @@ Uint32 DcmZLibOutputFilter::fillInputBuffer(const void *buf, Uint32 buflen)
       Uint32 len = DCMZLIBOUTPUTFILTER_BUFSIZE - inputBufCount_;
       if (len > buflen) len = buflen;
 
-      memcpy(inputBuf_ + (inputBufStart_ + inputBufCount_ - DCMZLIBOUTPUTFILTER_BUFSIZE), data, len);
+      memcpy(inputBuf_ + (inputBufStart_ + inputBufCount_ - DCMZLIBOUTPUTFILTER_BUFSIZE), data, (size_t)len);
 
       inputBufCount_ += len;
       result += len;
@@ -369,7 +369,11 @@ void dcostrmz_dummy_function()
 /*
  * CVS/RCS Log:
  * $Log: dcostrmz.cc,v $
- * Revision 1.2  2002-08-29 15:57:49  meichel
+ * Revision 1.3  2002-12-20 14:55:34  wilkens
+ * Inserted three casts in order to get rid of compiler warning on Solaris 2.5.1
+ * using compiler SC 2.0.1.
+ *
+ * Revision 1.2  2002/08/29 15:57:49  meichel
  * Updated zlib-related classes to correctly compile when WITH_ZLIB is undefined
  *
  * Revision 1.1  2002/08/27 16:55:54  meichel
